third_party/WebKit/LayoutTests/inspector/tracing/timeline-style-recalc-with-invalidator-invalidations.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>
function changeStylesAndDisplay(callback)
{
    requestAnimationFrame(function() {
        document.getElementById("testElementOne").className = "red";
        document.getElementById("testElementTwo").className = "red";
        var forceStyleRecalc = document.body.offsetTop;
        if (window.testRunner)
            testRunner.displayAsyncThen(callback);
    });
}

function changeMultipleStylesAndDisplay(callback)
{
    requestAnimationFrame(function() {
        var elementOne = document.getElementById("testElementOne");
        var elementTwo = document.getElementById("testElementTwo");
        var elementThree = document.getElementById("testElementThree");
        elementOne.className = "green";
        var forceStyleRecalc1 = document.body.offsetTop;
        elementOne.className = "blue";
        elementTwo.className = "blue";
        var forceStyleRecalc2 = document.body.offsetTop;
        elementOne.className = "snow";
        elementTwo.className = "snow";
        elementThree.className = "snow";
        var forceStyleRecalc3 = document.body.offsetTop;
        if (window.testRunner)
            testRunner.displayAsyncThen(callback);
    });
}

function changeMultipleSubframeStylesAndDisplay(callback)
{
    requestAnimationFrame(function() {
        var innerDocument = frames[0].document;
        var elementOne = innerDocument.getElementById("testElementOne");
        var elementTwo = innerDocument.getElementById("testElementTwo");
        var elementThree = innerDocument.getElementById("testElementThree");
        elementOne.className = "green";
        var forceStyleRecalc1 = innerDocument.body.offsetTop;
        elementOne.className = "blue";
        elementTwo.className = "blue";
        var forceStyleRecalc2 = innerDocument.body.offsetTop;
        elementOne.className = "snow";
        elementTwo.className = "snow";
        elementThree.className = "snow";
        var forceStyleRecalc3 = innerDocument.body.offsetTop;
        if (window.testRunner)
            testRunner.displayAsyncThen(callback);
    });
}

function test()
{
    var currentPanel = WebInspector.inspectorView.currentPanel();
    InspectorTest.assertEquals(currentPanel._panelName, "timeline", "Current panel should be the timeline.");
    Runtime.experiments.enableForTest("timelineInvalidationTracking");

    InspectorTest.runTestSuite([
        function testLocalFrame(next)
        {
            InspectorTest.invokeAsyncWithTimeline("changeStylesAndDisplay", function() {
                var record = InspectorTest.findFirstTimelineRecord(WebInspector.TimelineModel.RecordType.RecalculateStyles);
                var invalidations = record._event.invalidationTrackingEvents;
                InspectorTest.assertEquals(invalidations.length, 2);
                InspectorTest.assertEquals(invalidations[0].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(invalidations[0].nodeName, "DIV id='testElementOne' class='red'");
                InspectorTest.assertEquals(invalidations[0].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(invalidations[0].cause.stackTrace.length, 1);
                InspectorTest.assertEquals(invalidations[1].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(invalidations[1].nodeName, "DIV id='testElementTwo' class='red'");
                InspectorTest.assertEquals(invalidations[1].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(invalidations[1].cause.stackTrace.length, 1);
                next();
            });
        },

        function multipleStyleRecalcs(next)
        {
            InspectorTest.invokeAsyncWithTimeline("changeMultipleStylesAndDisplay", function() {
                var firstRecord = InspectorTest.findTimelineRecord(WebInspector.TimelineModel.RecordType.RecalculateStyles, 0);
                var firstInvalidations = firstRecord._event.invalidationTrackingEvents;
                InspectorTest.assertEquals(firstInvalidations.length, 1);
                InspectorTest.assertEquals(firstInvalidations[0].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(firstInvalidations[0].nodeName, "DIV id='testElementOne' class='green'");
                InspectorTest.assertEquals(firstInvalidations[0].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(firstInvalidations[0].cause.stackTrace.length, 1);

                var secondRecord = InspectorTest.findTimelineRecord(WebInspector.TimelineModel.RecordType.RecalculateStyles, 1);
                var secondInvalidations = secondRecord._event.invalidationTrackingEvents;
                InspectorTest.assertEquals(secondInvalidations.length, 2);
                InspectorTest.assertEquals(secondInvalidations[0].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(secondInvalidations[0].nodeName, "DIV id='testElementOne' class='blue'");
                InspectorTest.assertEquals(secondInvalidations[0].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(secondInvalidations[0].cause.stackTrace.length, 1);
                InspectorTest.assertEquals(secondInvalidations[1].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(secondInvalidations[1].nodeName, "DIV id='testElementTwo' class='blue'");
                InspectorTest.assertEquals(secondInvalidations[1].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(secondInvalidations[1].cause.stackTrace.length, 1);

                var thirdRecord = InspectorTest.findTimelineRecord(WebInspector.TimelineModel.RecordType.RecalculateStyles, 2);
                var thirdInvalidations = thirdRecord._event.invalidationTrackingEvents;
                InspectorTest.assertEquals(thirdInvalidations.length, 3);
                InspectorTest.assertEquals(thirdInvalidations[0].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(thirdInvalidations[0].nodeName, "DIV id='testElementOne' class='snow'");
                InspectorTest.assertEquals(thirdInvalidations[0].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(thirdInvalidations[0].cause.stackTrace.length, 1);
                InspectorTest.assertEquals(thirdInvalidations[1].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(thirdInvalidations[1].nodeName, "DIV id='testElementTwo' class='snow'");
                InspectorTest.assertEquals(thirdInvalidations[1].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(thirdInvalidations[1].cause.stackTrace.length, 1);
                InspectorTest.assertEquals(thirdInvalidations[2].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(thirdInvalidations[2].nodeName, "DIV id='testElementThree' class='snow'");
                InspectorTest.assertEquals(thirdInvalidations[2].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(thirdInvalidations[2].cause.stackTrace.length, 1);
                next();
            });
        },

        function testSubframe(next)
        {
            InspectorTest.invokeAsyncWithTimeline("changeMultipleSubframeStylesAndDisplay", function() {
                var firstRecord = InspectorTest.findTimelineRecord(WebInspector.TimelineModel.RecordType.RecalculateStyles, 0);
                var firstInvalidations = firstRecord._event.invalidationTrackingEvents;
                InspectorTest.assertEquals(firstInvalidations.length, 1);
                InspectorTest.assertEquals(firstInvalidations[0].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(firstInvalidations[0].nodeName, "DIV id='testElementOne' class='green'");
                InspectorTest.assertEquals(firstInvalidations[0].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(firstInvalidations[0].cause.stackTrace.length, 1);

                var secondRecord = InspectorTest.findTimelineRecord(WebInspector.TimelineModel.RecordType.RecalculateStyles, 1);
                var secondInvalidations = secondRecord._event.invalidationTrackingEvents;
                InspectorTest.assertEquals(secondInvalidations.length, 2);
                InspectorTest.assertEquals(secondInvalidations[0].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(secondInvalidations[0].nodeName, "DIV id='testElementOne' class='blue'");
                InspectorTest.assertEquals(secondInvalidations[0].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(secondInvalidations[0].cause.stackTrace.length, 1);
                InspectorTest.assertEquals(secondInvalidations[1].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(secondInvalidations[1].nodeName, "DIV id='testElementTwo' class='blue'");
                InspectorTest.assertEquals(secondInvalidations[1].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(secondInvalidations[1].cause.stackTrace.length, 1);

                var thirdRecord = InspectorTest.findTimelineRecord(WebInspector.TimelineModel.RecordType.RecalculateStyles, 2);
                var thirdInvalidations = thirdRecord._event.invalidationTrackingEvents;
                InspectorTest.assertEquals(thirdInvalidations.length, 3);
                InspectorTest.assertEquals(thirdInvalidations[0].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(thirdInvalidations[0].nodeName, "DIV id='testElementOne' class='snow'");
                InspectorTest.assertEquals(thirdInvalidations[0].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(thirdInvalidations[0].cause.stackTrace.length, 1);
                InspectorTest.assertEquals(thirdInvalidations[1].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(thirdInvalidations[1].nodeName, "DIV id='testElementTwo' class='snow'");
                InspectorTest.assertEquals(thirdInvalidations[1].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(thirdInvalidations[1].cause.stackTrace.length, 1);
                InspectorTest.assertEquals(thirdInvalidations[2].type, WebInspector.TimelineModel.RecordType.StyleRecalcInvalidationTracking);
                InspectorTest.assertEquals(thirdInvalidations[2].nodeName, "DIV id='testElementThree' class='snow'");
                InspectorTest.assertEquals(thirdInvalidations[2].cause.reason, "StyleInvalidator");
                InspectorTest.assertGreaterOrEqual(thirdInvalidations[2].cause.stackTrace.length, 1);
                next();
            });
        }
    ]);
}
</script>
<style>
    .testHolder > .red { background-color: red; }
    .testHolder > .green { background-color: green; }
    .testHolder > .blue { background-color: blue; }
    .testHolder > .snow { background-color: snow; }
</style>
</style>
</head>
<body onload="runTest()">
<p>Tests the Timeline API instrumentation of style recalc events with invalidations.</p>
<div class="testHolder">
<div id="testElementOne">PASS</div><div id="testElementTwo">PASS</div><div id="testElementThree">PASS</div>
</div>
<iframe src="resources/timeline-iframe-with-style.html" style="position: absolute; left: 40px; top: 40px; width: 100px; height: 100px; border: none"></iframe>
</body>
</html>
